+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.7] - 2024-04-10
 
+### Added
+- [config.featureBlackList] z2m features with name in the list will be ignored (e.g. device_temperature) (https://github.com/nitaybz).
+- [config.switchList] Device with friendly name in the list will be exposed like switch.
+- [config.lightList] Device with friendly name in the list will be exposed like light.
+- [config.outletList] Devices with friendly name in the list will be exposed like outlet.
+
 ### Fixed
 
-- [payload]: Fixed the case when z2m has some issues and send null or undefined in the payload (https://github.com/khaidakin).
+- [payload]: Fixed the case when z2m sends null or undefined in the payload (https://github.com/khaidakin).
 
 ## [2.0.6] - 2024-04-09